BakkaLinux: Added EXT2 rootfs support
To generate an ext2 root filesystem:
make FSTYPE=rootfs.ext2

In addition to the normal build process, this will copy
busybox binaries and kernel modules to a new directory, ./fs.
The contents of ./root is also copied to ./fs.
Then a raw filesystem image is created based on the contents
of ./fs.

External program 'genext2fs' is required.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>